Implement multi-repo support
The YAML config can now contain a list of repos in the `repos` key.
If this key is not found the parser falls back to the old behavior and
interpretes the file as the configuration for one repo.

The backup and create-svg commands now take a parameter which defines
the repo from the config file which is to be backed up.

The scan command writes the multi-repo format.

The show and write-config commands handle a list of repos now.